portal/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
sqlmodel==0.0.22
paramiko==3.5.0
pyyaml==6.0.2
python-multipart==0.0.20
websockets==13.1
argon2-cffi==23.1.0
webauthn==2.4.0
itsdangerous==2.2.0
